requirements.txt
pyspark>=3.0.0
python-jobspy>=1.1.76
pandas>=1.5.0
pytest>=7.0.0